requirements.txt
git+https://github.com/halcy/Mastodon.py.git
Mastodon.py
feedparser
python-dateutil
